backend/requirements.txt
asgiref==3.7.2
Django==4.2.6
sqlparse==0.4.4
typing_extensions==4.8.0
environs
djangorestframework
django-cors-headers==4.3.0
djoser
